- PSR-2 를 맞추세요. - 굳이 서비스로 리펙토링하는 시점은 그럴필요가 있을때 하세요. - 코드를 이쁘게 정리했으면 합니다. 커밋하기전에 코드정리를 하세요. - 파사드는 되도록이면 네임스페이스를 명시해주면 자동완성이 됩니다. - class 프로퍼티는 type 지정을 해주세요.
